Use ET_DYN instead of ET_REL modules
ET_DYN also known as .so files are actually meant to be loaded dynamically.
Most of the work of linking is done by normal linker and we need to do
only simple relocation by offset. This significantly simplifies our
dynamic loading. Moreover the tools are meant to produce .so files and so
it's easier to get them to produce correct output. Also it reduce the
core size on i386-pc by 44 bytes on config biosdisk+part_msdos+ext2

This required to use -fPIC on mips as otherwise linker chokes on relocations
incompatible with -shared

This was originally meant for Rust as rustc produces .so output
as a possibility but it simplifies the code, so it's useful independently of
Rust

Signed-off-by: Vladimir Serbinenko <phcoder@gmail.com>